requirements.txt
bcrypt==4.1.1
mysql-connector==2.2.9
websockets==12.0
wheel==0.42.0
